# backend/src/Api/SharpPizza.Api/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SharpPizza.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Payment": {
      "post": {
        "tags": [
          "Payment"
        ],
        "operationId": "CreatePayment",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreatePaymentCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OrderVm"
            }
          }
        }
      }
    },
...
    "/api/Usuario/login": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "Login",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
    "/api/Usuario/register": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "Register",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Nombre",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Apellido",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Telefono",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Foto",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "FotoUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FotoId",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Password",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Username",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
    "/api/Usuario/forgotpassword": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "ForgotPassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendPasswordCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Usuario/resetpassword": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "ResetPassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordByTokenCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Usuario/updatepassword": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "UpdatePassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        }
      }
    },
    "/api/Usuario/update": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "operationId": "Update",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Nombre",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Apellido",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Telefono",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Foto",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "FotoUrl",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FotoId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
    "/api/Usuario/updateAdminUser": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "operationId": "UpdateAdminUser",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAdminUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        }
      }
    },
    "/api/Usuario/updateAdminStastusUser": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "operationId": "UpdateAdminStastusUser",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAdminStatusUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        }
      }
    },
    "/api/Usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "operationId": "GetUsuarioById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
...
    "/api/Usuario/username/{username}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "operationId": "GetUsuarioByUsername",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
    "/api/Usuario/paginationAdmin": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "operationId": "PaginationUser",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Search",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "Sort",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageIndex",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioPaginationVm"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddressVm": {
      "type": "object",
      "properties": {
        "direccion": {
          "type": "string"
        },
        "ciudad": {
          "type": "string"
        },
        "departamento": {
          "type": "string"
        },
        "codigoPostal": {
          "type": "string"
        },
        "pais": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "direccionEnvio": {
          "$ref": "#/definitions/AddressVm"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CreatePaymentCommand": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "shoppingCartMasterId": {
          "format": "uuid",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "LoginUserCommand": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderItemVm": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "productItemId": {
          "format": "int32",
          "type": "integer"
        },
        "productNombre": {
          "type": "string"
        },
        "imagenUrl": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "OrderVm": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderAddress": {
          "$ref": "#/definitions/AddressVm"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItemVm"
          }
        },
        "subtotal": {
          "format": "double",
          "type": "number"
        },
        "impuesto": {
          "format": "double",
          "type": "number"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "precioEnvio": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "paymentIntentId": {
          "type": "string"
        },
        "clientSecret": {
          "type": "string"
        },
        "stripeApiKey": {
          "type": "string"
        },
        "compradorUsername": {
          "type": "string"
        },
        "compradorNombre": {
          "type": "string"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "stastusLabel": {
          "type": "string",
          "readOnly": true
        }
      },
      "additionalProperties": false
    },
...
    "ResetPasswordByTokenCommand": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ResetPasswordCommand": {
      "type": "object",
      "properties": {
        "newPassword": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "SendPasswordCommand": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Unit": {
      "type": "object",
      "additionalProperties": false
    },
    "UpdateAdminStatusUserCommand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UpdateAdminUserCommand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "UsuarioPaginationVm": {
      "type": "object",
      "properties": {
        "count": {
          "format": "int32",
          "type": "integer"
        },
        "pageIndex": {
          "format": "int32",
          "type": "integer"
        },
        "pageSize": {
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Usuario"
          }
        },
        "pageCount": {
          "format": "int32",
          "type": "integer"
        },
        "resultByPage": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    }
  }
}